local device, device discovery, and service search are almost done, but other parts not yet.
some known problems in the code.

1 - cancelInquiry() issues an hci command to stop device inquiry, it will work well except in this case:
if there is another program that uses the bluetooth device in the same time and the other program has started an inquiry, then our program started an inquiry before the previous inquiry ends, a call to cancelInquiry() will stop the inquiry of the other program not ours (if it hasn't been ended yet). 

a probable solution to this is to check if there is an inquiry already running before I start my own, but till now I don't know if this possible or not and asking bluez community about this, results in a simple answer "don't use these APIs, use DBus". 
also note that avetana has no implementation of DiscoveryAgent.cancelInquiry().


2 - cancelServiceSearch() is implemented as "return false" and I don't know how to stop a service search.

3 - In populating service records,
calling sdp_service_attr_req() causes a crash if the attrid_list contains a big number of entries. (I guess it is a memory problem).